+++ tsconfig.json
@@ -21,7 +21,8 @@
       "@/*": ["./*"],
       "@/components/*": ["./app/components/*"],
       "@/utils/*": ["./app/utils/*"],
-      "@/definitions/*": ["./app/definitions/*"]
+      "@/definitions/*": ["./app/utils/definitions/*"],
+      "@/actions/*": ["./app/utils/actions/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
